Time Seriess Classifier

Allows training and predicting on time-series classification data. Internally writes the data to an .arff file and reads it from the command line in the Java Virtual Machine. In some cases, the memory of the JVM is not sufficient. In this case, the memory for the JVM can be set e.g. to 2 GB by setting options(java.options = "-Xmx2048m") before loading the package.

TSClassifier

Format

An object of class R6ClassGenerator of length 24.

Details

Expects the data to either be a data.frame or a character path to a data set in the '.arff' format. In case a 'data.frame' is provided, the data is saved to '.$model_path' and then read into the JVM from there. The target variable's name is assumed to be 'target'.

Methods

new(classifier, model_path): Initialize the class.

train(data, target, par_vals, data_path): Delegates to train_tsc.

predict(newdata): Delegates to predict_tsc.

resample(data, target, par_vals, data_path): Delegates to resample_tsc.

cleanup(): Remove saved model files.

Fields

classifier: character(1)
Classifier to use, see ?tsc_classifiers for a list of available classifiers.

target: character(1)
Target variable.

data: character(1)|data.frame
Either a path to the dataset or a data.frame that should be saved to disk for modeling. In case a data.frame is provided, the dataset is saved to disk via data_to_path.

newdata: character(1)|data.frame
See data.

par_vals: list(1)
Named list of hyperparamter_values.

data_path: character(1)
Save path for the data.

target_levels: character(n)
Levels of the target variable, if a data.frame is supplied.

model_path: character(1)
Path to store the resulting model to. Default NULL creates and stores to a temporary file.

trained: logical(1)
Was the model trained?

Examples

  data = data.frame(matrix(rnorm(300), nrow = 30))
  data$class = factor(sample(letters[1:2], 10, replace = TRUE))
  tsc = TSClassifier$new("timeseriesweka.classifiers.BOSS")
  tsc$train(data[1:15,], target = "class", par_vals = list(setMaxEnsembleSize = 1))
  p = tsc$predict(data[10:20,])