Return available classifiers

Run tsc_classifiers() to obtain available classifiers

tsc_classifiers()

Value

character Names of available classifiers.

Details

The following classifiers are available:

Ensemble Classifiers

timeseriesweka.classifiers.ElasticEnsemble
Combination of nearest Neighbour (NN) classifiers that use elastic distance measures
Hyperparameters: None

timeseriesweka.classifiers.FlatCote
Collective of Transformation Ensembles (Bagnall et al.,2015)
Hyperparameters: None

Base-learners of ElasticEnsemble:

timeseriesweka.classifiers.ensembles.elastic_ensemble.WDTW1NN
Elastic Ensemble of Nearest Neighbour Algorithms:
Weighted Dynamic Time Warping 1 Nearest Neighbour
Hyperparameters: None

timeseriesweka.classifiers.ensembles.elastic_ensemble.ED1NN
Euclidean distance with 1 nearest neighbor
Hyperparameters: None

timeseriesweka.classifiers.ensembles.elastic_ensemble.DTW1NN
Dynamic time warping with 1 nearest neighbor
Hyperparameters:

setWindow: double range: [1, Inf]

timeseriesweka.classifiers.ensembles.elastic_ensemble.ERP1NN
edit distance with real penalty with 1 nearest neighbor
Hyperparameters: None

timeseriesweka.classifiers.ensembles.elastic_ensemble.LCSS1NN
longest common subsequence with 1 nearest neighbor
Hyperparameters: None

timeseriesweka.classifiers.ensembles.elastic_ensemble.TWE1NN
Time Warp Edit with 1 nearest neighbor
Hyperparameters: None

timeseriesweka.classifiers.ensembles.elastic_ensemble.MSM1NN
Move-Split-Merge with 1 nearest neighbor
Hyperparameters: None

Differential Distance Based Classifiers

timeseriesweka.classifiers.NN_CID
Complexity Invariant distance with k nearest neighbor
Hyperparameters: None

timeseriesweka.classifiers.DD_DTW
Derivative dynamic time warping
Hyperparameters: None

timeseriesweka.classifiers.DTD_C
Derivative transform distance
Hyperparameters: None

Dictionary based Classifiers

timeseriesweka.classifiers.BOSS
Bag of SFA Symbols
Hyperparameters:

setMaxEnsembleSize: integer(1) range: [1, Inf]

timeseriesweka.classifiers.BagOfPatterns
Bag of Patterns
Hyperparameters: None

timeseriesweka.classifiers.SAX_1NN
Symbolic Aggregate Approximation
Hyperparameters: None

timeseriesweka.classifiers.SAXVSM
Symbolic Aggregate Approximation - Vector Space Model
Hyperparameters: None

Shapelet based Classifiers

timeseriesweka.classifiers.ShapeletTransformClassifier
Shapelet transformation that separates the Shapelet discovery from the classifier by finding the top k Shapelets in a single run
Hyperparameters:

setTransformType: character(1)
values: "univariate","uni","shapeletd","shapeleti"

setNumberOfShapelets: integer(1) range: [1, Inf]

timeseriesweka.classifiers.FastShapelets
Fast Shapelets (FS)
Hyperparameters: None

timeseriesweka.classifiers.LearnShapelets
Learned Shapelets (LS):
Hyperparameters: None

Interval based Classifiers

timeseriesweka.classifiers.TSF
Time Series Forest (Deng et al.,2013)
Hyperparameters:

setNumTrees: integer(1) range: [1, Inf]

timeseriesweka.classifiers.TSBF
Time Series Bag of Features (TSBF)
Hyperparameters:

setZLevel: double(1)

timeseriesweka.classifiers.LPS
Learned Pattern Similarity (LPS)
Hyperparameters: None

Time Series Classifier

timeseriesweka.classifiers.DTW_kNN
specialization of kNN that can only be used with the efficient DTW distance
Hyperparameters:

setMaxR: double(1) range: \[0, 1] set max window size

timeseriesweka.classifiers.FastDTW_1NN
fast Dynamic time warping with 1 nearest neighbor
Hyperparameters:

setR: double(1)

timeseriesweka.classifiers.SlowDTW_1NN
compare with FastDTW_1NN
Hyperparameters:

setR: double(1)

Weka Classifiers

Several WEKA classifiers have been implemented in the Time-Series Classification Bake-off.
The use of those classifiers is discouraged from within TSClassification, but nonetheless implemented for completeness.
We advise to use the official implementations from package RWeka (https://cran.r-project.org/web/packages/RWeka/index.html) for greater flexibility and improved support for setting hyperparameters.

weka.classifiers.functions.Logistic

weka.classifiers.bayes.BayesNet

weka.classifiers.bayes.NaiveBayes

weka.classifiers.functions.Logistic

weka.classifiers.functions.MultilayerPerceptron

weka.classifiers.functions.SMO

weka.classifiers.meta.RotationForest

weka.classifiers.trees.J48

weka.classifiers.trees.RandomForest

Examples

tsc_classifiers()

#>  [1] "timeseriesweka.classifiers.ensembles.elastic_ensemble.ED1NN"  
#>  [2] "timeseriesweka.classifiers.ensembles.elastic_ensemble.DTW1NN" 
#>  [3] "timeseriesweka.classifiers.ensembles.elastic_ensemble.ERP1NN" 
#>  [4] "timeseriesweka.classifiers.ensembles.elastic_ensemble.LCSS1NN"
#>  [5] "timeseriesweka.classifiers.ensembles.elastic_ensemble.WDTW1NN"
#>  [6] "timeseriesweka.classifiers.ensembles.elastic_ensemble.TWE1NN" 
#>  [7] "timeseriesweka.classifiers.ensembles.elastic_ensemble.MSM1NN" 
#>  [8] "timeseriesweka.classifiers.ElasticEnsemble"                   
#>  [9] "timeseriesweka.classifiers.FlatCote"                          
#> [10] "timeseriesweka.classifiers.NN_CID"                            
#> [11] "timeseriesweka.classifiers.DD_DTW"                            
#> [12] "timeseriesweka.classifiers.DTD_C"                             
#> [13] "timeseriesweka.classifiers.BagOfPatterns"                     
#> [14] "timeseriesweka.classifiers.SAX_1NN"                           
#> [15] "timeseriesweka.classifiers.SAXVSM"                            
#> [16] "timeseriesweka.classifiers.BOSS"                              
#> [17] "timeseriesweka.classifiers.FastShapelets"                     
#> [18] "timeseriesweka.classifiers.ShapeletTransformClassifier"       
#> [19] "timeseriesweka.classifiers.LearnShapelets"                    
#> [20] "timeseriesweka.classifiers.TSF"                               
#> [21] "timeseriesweka.classifiers.TSBF"                              
#> [22] "timeseriesweka.classifiers.LPS"                               
#> [23] "weka.classifiers.bayes.BayesNet"                              
#> [24] "weka.classifiers.bayes.NaiveBayes"                            
#> [25] "weka.classifiers.functions.Logistic"                          
#> [26] "weka.classifiers.functions.MultilayerPerceptron"              
#> [27] "weka.classifiers.functions.SMO"                               
#> [28] "weka.classifiers.meta.RotationForest"                         
#> [29] "weka.classifiers.trees.J48"                                   
#> [30] "weka.classifiers.trees.RandomForest"                          
#> [31] "timeseriesweka.classifiers.DTW_kNN"                           
#> [32] "timeseriesweka.classifiers.FastDTW_1NN"                       
#> [33] "timeseriesweka.classifiers.RISE"                              
#> [34] "timeseriesweka.classifiers.SlowDTW_1NN"